# openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "BISQ Support",
        "version": "0.1.0",
        "description": "BISQ Support API - A RAG-based support assistant with admin authentication.\n\n## Security\n\n### Transport Security (CKV_OPENAPI_20)\nProduction deployments use HTTPS exclusively via nginx reverse proxy configuration.\nAPI keys are only transmitted over secure channels in production.\nLocal development uses HTTP for debugging convenience.\n\n### Authentication Design (CKV_OPENAPI_4, CKV_OPENAPI_5)\nThis API uses **per-endpoint security** by design:\n- Public endpoints: /health, /chat, /feedback (no authentication)\n- Admin endpoints: /admin/* (require AdminApiKeyAuth header)\n- Exception: /admin/auth/logout is public (no authentication required)\n\nNo global security is defined because public and authenticated endpoints coexist.\n\n### Array Limits (CKV_OPENAPI_21)\nArray properties intentionally omit maxItems to allow flexible data volumes.\nActual limits are enforced by:\n- Application logic (e.g., FeedbackRequest.conversation_history maxItems=50)\n- Backend validation in FastAPI route handlers\n- Database constraints and query limits"
    },
    "paths": {
...
        "/feedback/submit": {
            "post": {
                "tags": [
                    "Feedback"
                ],
                "summary": "Submit Feedback",
                "description": "Submit feedback for a chat response.",
                "operationId": "submit_feedback_feedback_submit_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/FeedbackRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/feedback/explanation": {
            "post": {
                "tags": [
                    "Feedback"
                ],
                "summary": "Submit Feedback Explanation",
                "description": "Submit explanation for negative feedback.\n\nThis endpoint receives explanations about why an answer was unhelpful.\nIt updates the existing feedback with the explanation and categorizes issues.",
                "operationId": "submit_feedback_explanation_feedback_explanation_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "additionalProperties": true,
                                "type": "object",
                                "title": "Explanation Data"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "additionalProperties": true,
                                    "type": "object",
                                    "title": "Response Submit Feedback Explanation Feedback Explanation Post"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ConversationMessage": {
                "properties": {
                    "role": {
                        "type": "string",
                        "enum": [
                            "user",
                            "assistant"
                        ],
                        "title": "Role",
                        "description": "Message role"
                    },
                    "content": {
                        "type": "string",
                        "maxLength": 10000,
                        "title": "Content",
                        "description": "Message content (max 10KB)"
                    }
                },
                "type": "object",
                "required": [
                    "role",
                    "content"
                ],
                "title": "ConversationMessage",
                "description": "Individual message in a conversation."
            },
...
            "FeedbackRequest": {
                "properties": {
                    "message_id": {
                        "type": "string",
                        "pattern": "^[a-f0-9-]{36}$",
                        "title": "Message Id",
                        "description": "UUID of the message"
                    },
                    "question": {
                        "type": "string",
                        "maxLength": 5000,
                        "title": "Question",
                        "description": "User question"
                    },
                    "answer": {
                        "type": "string",
                        "maxLength": 20000,
                        "title": "Answer",
                        "description": "Assistant answer"
                    },
                    "rating": {
                        "type": "integer",
                        "maximum": 1.0,
                        "minimum": 0.0,
                        "title": "Rating",
                        "description": "0 for negative, 1 for positive"
                    },
                    "explanation": {
                        "anyOf": [
                            {
                                "type": "string",
                                "maxLength": 5000
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Explanation",
                        "description": "Feedback explanation"
                    },
                    "conversation_history": {
                        "anyOf": [
                            {
                                "items": {
                                    "$ref": "#/components/schemas/ConversationMessage"
                                },
                                "type": "array",
                                "maxItems": 50
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Conversation History",
                        "description": "Conversation context (max 50 messages)"
                    }
                },
                "type": "object",
                "required": [
                    "message_id",
                    "question",
                    "answer",
                    "rating"
                ],
                "title": "FeedbackRequest",
                "description": "Request model for submitting feedback."
            },
...
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            }
        },
        "securitySchemes": {
            "AdminApiKeyAuth": {
                "type": "apiKey",
                "in": "header",
                "name": "Authorization",
                "description": "Enter the token with the `Bearer ` prefix, e.g. `Bearer abcdef12345`"
            }
        }
    }
}